pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claim-photo-editor"
version = "0.1.0"
description = "A desktop application for captioning photos and generating PDFs"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "James Curtin" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "PySide6>=6.8.0",
    "Pillow>=11.0.0",
    "piexif>=1.1.3",
    "reportlab>=4.2.0",
    "requests>=2.32.0",
    "packaging>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-qt>=4.4.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-requests>=2.32.0",
    "types-Pillow>=10.2.0",
]

[project.scripts]
claim-photo-editor = "claim_photo_editor.main:main"

[project.gui-scripts]
claim-photo-editor-gui = "claim_photo_editor.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/claim_photo_editor"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "PLR0913",  # Too many arguments - common in UI code
    "PLR2004",  # Magic value comparison - too strict
]

[tool.ruff.lint.isort]
known-first-party = ["claim_photo_editor"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "piexif.*",
    "reportlab.*",
    "PySide6.*",
    "PIL.*",
    "requests.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "claim_photo_editor.services.updater",
    "claim_photo_editor.services.pdf_generator",
]
disable_error_code = ["import-untyped", "import-not-found"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=claim_photo_editor --cov-report=term-missing --cov-report=html"
qt_api = "pyside6"

[tool.coverage.run]
source = ["src/claim_photo_editor"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
